Feature/iris/add button to ingest one lecture (#8796)
Add Johannes review + Add a button to ingest one lecture only

+++ src/main/java/de/tum/in/www1/artemis/service/connectors/pyris/dto/lectureingestionwebhook/PyrisLectureUnitWebhookDTO.java
@@ -8,6 +8,6 @@ import com.fasterxml.jackson.annotation.JsonInclude;
  * providing necessary details such as lecture and course identifiers, names, and descriptions.
  */
 @JsonInclude(JsonInclude.Include.NON_EMPTY)
-public record PyrisLectureUnitWebhookDTO(Boolean toUpdate, String artemisBaseUrl, String pdfFile, int lectureUnitId, String lectureUnitName, int lectureId, String lectureName,
-        int courseId, String courseName, String courseDescription) {
+public record PyrisLectureUnitWebhookDTO(Boolean toUpdate, String artemisBaseUrl, String pdfFile, Long lectureUnitId, String lectureUnitName, Long lectureId, String lectureName,
+        Long courseId, String courseName, String courseDescription) {
 }
